[review] Extract routing resolution and model invocation from ChatModelAction
Per wenjin272's review: ChatModelAction had grown to ~1000 lines across
several responsibilities. Routing-cohesive pieces move to package-private
classes:

- ResolvedModelRoute: the resolution outcome (candidates, fallback policy,
  decision facts) plus attempt order, routed durable-call ids, and the
  model_routing response block — the former RoutingSelection inner class
  and its free-function helpers, now methods on the type they describe.
- ModelRoutingResolver: strategy execution inside the durable route call,
  decision normalization, and ModelRoutingEvent emission.
- ChatModelInvoker: one-candidate invocation with durable-call + retry
  machinery (chatWithRetries, ChatAttemptResult, ChatAttemptFailed).

ChatModelAction keeps event orchestration, the fallback loop, tool-loop
state, and retry-stats accounting. recordChatTokenMetrics and
generateStructuredOutput stay put to avoid churning code (and tests)
that main owns. No behavior change; full plan suite green.
